+++ packages/editor/src/store/actions.js
@@ -11,6 +11,7 @@ import {
 	getDefaultBlockName,
 	createBlock,
 } from '@wordpress/blocks';
+import deprecated from '@wordpress/deprecated';
 
 /**
  * Returns an action object used in signalling that editor has initialized with
@@ -373,6 +374,14 @@ export function setTemplateValidity( isValid ) {
  * @return {Object} Action object.
  */
 export function checkTemplateValidity() {
+	// TODO: Hello future deprecation remover. Please ensure also to remove all
+	// references to CHECK_TEMPLATE_VALIDITY, notably its effect handler.
+	deprecated( 'checkTemplateValidity action (`core/editor`)', {
+		version: '4.1',
+		plugin: 'Gutenberg',
+		hint: 'Validity is verified automatically upon block reset.',
+	} );
+
 	return {
 		type: 'CHECK_TEMPLATE_VALIDITY',
 	};